requirements.txt
python-dotenv==1.1.0
numpy==1.26.4
pandas==2.2.2
matplotlib==3.10.5
seaborn==0.13.2
pydantic==2.9.2
pytest==8.3.3
yfinance==0.2.59
beautifultable==1.1.0